requirements.txt
paho-mqtt==1.5.1
PyYAML==5.4.1
wakeonlan==2.0.1
